[notify] dwmblocks: removed (upstream transition)
Our `dwmblocks`-fork was renamed to `rawstat` and moved from
Zeppe-Lin Org.

This decision stems from a shift in focus and interest.
To avoid any potential misunderstanding or misconduct regarding the
maintenance and direction of this particular `dwmblocks`-fork, I
(sighook) decided to separate it from the Zeppe-Lin Org.

What does this means for you? If you are currently using the `dwmblocks`
package from Zeppe-Lin, it will no longer receive updates or be
available for installation from our official pkgsrc repositories.

For those who appreciate and wish to continue using this specified fork,
it has been renamed to `rawstat` and has found a new home at:
<https://github.com/rawmkit/rawstat>.  This new repository will be where
any future development or maintenance of this particular window manager
will occur.

I encourage interested users to migrate to the `rawstat` project
directly, and (optionally) send a patch with new *package source* to
<zeppe-lin-dev@freelists.org> or open PR at
<https://github.com/zeppe-lin/pkgsrc-desktop>.